added checks

+++ utilities/create_obsidian_vault.sh
@@ -2,23 +2,36 @@
 source "$HOME/scripts/settings.sh"
 
 echo "creating folder $1"
-mkdir "$1"
+if [[ ! -d "$1" ]]; then
+    mkdir "$1"
+fi
 
 echo "create subfolders "
-mkdir "$1/pages"
-mkdir "$1/journals"
-mkdir "$1/assets"
-
+if [[ ! -d "$1/journals" ]]; then
+    mkdir "$1/journals"
+fi
+if [[ ! -d "$1/assets" ]]; then
+    mkdir "$1/assets"
+fi
+if [[ ! -d "$1/pages" ]]; then
+    mkdir "$1/pages"
+fi
+if [[ ! -d "$1/.git" ]]; then
+    echo "entering $1"
+    cd "$1"
+    git init
+fi
 echo "creating gitignore"
-echo '.obsidian/workspace.json'> "$1/.gitignore"
+if [[ ! -f "$1/.gitignore" ]]; then
+    echo '.obsidian/workspace.json' >"$1/.gitignore"
+    echo "commiting gitignore"
+    cd "$1"
+    git add .gitignore
+    git commit -m 'added gitignore'
 
-echo "entering $1"
-cd "$1"
-git init
+fi
 
-echo "commiting gitignore"
-git add .gitignore
-git commit -m 'added gitignore'
-
-echo "copying obsidian files"
-cp -r $SCRIPTS_HOME_FOLDER/utilities/obsidian-files ./.obsidian
+if [[ ! -d "$1/.obsidian" ]]; then
+    echo "copying obsidian files"
+    cp -r $SCRIPTS_HOME_FOLDER/utilities/obsidian-files ./.obsidian
+fi
